Elysium release

+++ src/game/Movement/WaypointMovementGenerator.cpp
@@ -318,7 +318,7 @@ bool FlightPathMovementGenerator::Update(Player &player, const uint32 &diff)
 {
     int32 pointId = player.movespline->currentPathIdx();
     // currentPathIdx returns lastIdx + 1 at arrive
-    while (i_currentNode < pointId)
+    while (static_cast <int32>(i_currentNode) < pointId)
     {
         DoEventIfAny(player, (*i_path)[i_currentNode], true);
         DoEventIfAny(player, (*i_path)[i_currentNode], false);
